Move TurnMove interface to its own file
